fix(ci): resolve -Werror build failures, MPI lifecycle errors, and CI workflow issues
Replace deprecated distributed_vector(size, num_ranks, rank) constructors
with context-based constructors across integration tests. Add MPI_Finalized
checks in backend_skip.hpp to prevent post-finalize MPI calls. Fix
sign-conversion warnings in mpi_topology, stress tests, and work-stealing
pool tests. Add setup-python action to pages and coverage workflows for
PEP 668 compliance on Ubuntu 24.04. Suppress unused variable warning in
stable_sort example. Create scripts/format_code.sh referenced by
format-check workflow.

+++ tests/unit/backend/test_work_stealing_pool.cpp
@@ -128,7 +128,7 @@ TEST(WorkStealingPoolTest, MultipleTasksCorrectness) {
     }
 
     for (int i = 0; i < N; ++i) {
-        EXPECT_EQ(futures[i].get(), i * i);
+        EXPECT_EQ(futures[static_cast<size_t>(i)].get(), i * i);
     }
 }
